feat: :sparkles: create basic routing structure
