--- src/App.tsx
@@ -1,7 +1,0 @@
-import { Dashboard } from "./components/dashboard/Dashboard/Dashboard";
-
-function App() {
-  return <Dashboard />;
-}
-
-export default App;
