+++ src/components/home/Home.tsx
@@ -1,0 +1,4 @@
+const Home = () => {
+  return <p> This is the home page</p>;
+};
+export { Home };
