+++ src/components/imageLibrary/ImageLibrary.tsx
@@ -1,0 +1,5 @@
+const ImageLibrary = () => {
+  return <p>List of Images</p>;
+};
+
+export { ImageLibrary };
