+++ src/main.tsx
@@ -1,18 +1,9 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App.tsx";
-import { ThemeProvider } from "@mui/material";
-import { BrowserRouter, Route, Routes } from "react-router";
-import theme from "./themes/theme.ts";
+import App from "./components/App";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <ThemeProvider theme={theme}>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<App />} />
-        </Routes>
-      </BrowserRouter>
-    </ThemeProvider>
+    <App />
   </StrictMode>,
 );
